+++ contracts/multihop/src/lib.rs
@@ -5,7 +5,7 @@ mod storage;
 mod utils;
 
 #[allow(clippy::too_many_arguments)]
-pub mod lp_contract {
+pub mod xyk_pool {
     // The import will code generate:
     // - A ContractClient type that can be used to invoke functions on the contract.
     // - Any types in the contract that were annotated with #[contracttype].
@@ -15,10 +15,7 @@ pub mod lp_contract {
 }
 
 #[allow(clippy::too_many_arguments)]
-pub mod stable_lp_contract {
-    // The import will code generate:
-    // - A ContractClient type that can be used to invoke functions on the contract.
-    // - Any types in the contract that were annotated with #[contracttype].
+pub mod stable_pool {
     soroban_sdk::contractimport!(
         file = "../../target/wasm32-unknown-unknown/release/phoenix_pool_stable.wasm"
     );