+++ contracts/multihop/src/tests/swap.rs
@@ -1,8 +1,9 @@
+use crate::factory_contract::PoolType;
 // FIXM: Disable Referral struct
 // use crate::lp_contract::Referral;
 use crate::storage::Swap;
 use crate::tests::setup::{
-    deploy_and_initialize_factory, deploy_and_initialize_lp, deploy_and_mint_tokens,
+    deploy_and_initialize_factory, deploy_and_initialize_pool, deploy_and_mint_tokens,
     deploy_multihop_contract, deploy_token_contract,
 };
 
@@ -33,7 +34,7 @@ fn swap_three_equal_pools_no_fees() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -42,8 +43,9 @@ fn swap_three_equal_pools_no_fees() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -52,8 +54,9 @@ fn swap_three_equal_pools_no_fees() {
         token3.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -62,6 +65,7 @@ fn swap_three_equal_pools_no_fees() {
         token4.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -91,7 +95,7 @@ fn swap_three_equal_pools_no_fees() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &operations, &None, &None, &50i128);
-    multihop.swap(&recipient, &operations, &None, &50i128);
+    multihop.swap(&recipient, &operations, &None, &50i128, &PoolType::Xyk);
 
     // 5. check if it goes according to plan
     assert_eq!(token1.balance(&recipient), 0i128);
@@ -117,7 +121,7 @@ fn swap_three_equal_pools_no_fees_referral_fee() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -126,8 +130,9 @@ fn swap_three_equal_pools_no_fees_referral_fee() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -136,8 +141,9 @@ fn swap_three_equal_pools_no_fees_referral_fee() {
         token3.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -146,6 +152,7 @@ fn swap_three_equal_pools_no_fees_referral_fee() {
         token4.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -187,7 +194,7 @@ fn swap_three_equal_pools_no_fees_referral_fee() {
     //     &None,
     //     &50i128,
     // );
-    multihop.swap(&recipient, &operations, &None, &50i128);
+    multihop.swap(&recipient, &operations, &None, &50i128, &PoolType::Xyk);
 
     // 5. check if it goes according to plan
     assert_eq!(token1.balance(&recipient), 0i128);
@@ -214,7 +221,7 @@ fn swap_single_pool_no_fees() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -223,6 +230,7 @@ fn swap_single_pool_no_fees() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -242,7 +250,7 @@ fn swap_single_pool_no_fees() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, /*&None,*/ &operations, &None, &None, &50i128);
-    multihop.swap(&recipient, &operations, &None, &1_000);
+    multihop.swap(&recipient, &operations, &None, &1_000, &PoolType::Xyk);
 
     // 5. check if it goes according to plan
     assert_eq!(token1.balance(&recipient), 4_000i128); // -1_000 token0
@@ -264,7 +272,7 @@ fn swap_should_fail_when_spread_exceeds_the_limit() {
 
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -273,6 +281,7 @@ fn swap_should_fail_when_spread_exceeds_the_limit() {
         token2.address.clone(),
         2_000_000,
         None,
+        PoolType::Xyk,
     );
 
     let multihop = deploy_multihop_contract(&env, admin.clone(), &factory_client.address);
@@ -289,7 +298,7 @@ fn swap_should_fail_when_spread_exceeds_the_limit() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &operations, &None, &Some(50), &50);
-    multihop.swap(&recipient, &operations, &Some(50), &50);
+    multihop.swap(&recipient, &operations, &Some(50), &50, &PoolType::Xyk);
 }
 
 #[test]
@@ -306,7 +315,7 @@ fn swap_single_pool_with_fees() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -315,6 +324,7 @@ fn swap_single_pool_with_fees() {
         token2.address.clone(),
         1_000_000,
         Some(2000),
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -334,7 +344,7 @@ fn swap_single_pool_with_fees() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &operations, &None, &None, &300i128);
-    multihop.swap(&recipient, &operations, &None, &300i128);
+    multihop.swap(&recipient, &operations, &None, &300i128, &PoolType::Xyk);
 
     // 5. check if it goes according to plan
     // 1000 tokens initially
@@ -361,7 +371,7 @@ fn swap_three_different_pools_no_fees() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -370,8 +380,9 @@ fn swap_three_different_pools_no_fees() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -380,8 +391,9 @@ fn swap_three_different_pools_no_fees() {
         token3.address.clone(),
         2_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -390,6 +402,7 @@ fn swap_three_different_pools_no_fees() {
         token4.address.clone(),
         3_000_000,
         None,
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -420,7 +433,7 @@ fn swap_three_different_pools_no_fees() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &operations, &None, &None, &5_000i128);
-    multihop.swap(&recipient, &operations, &None, &5_000i128);
+    multihop.swap(&recipient, &operations, &None, &5_000i128, &PoolType::Xyk);
 
     // 5. check if it goes according to plan
     assert_eq!(token1.balance(&recipient), 0i128);
@@ -448,7 +461,7 @@ fn swap_three_different_pools_with_fees() {
     // 1. deploy factory
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -457,8 +470,9 @@ fn swap_three_different_pools_with_fees() {
         token2.address.clone(),
         1_000_000,
         Some(1_000),
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -467,8 +481,9 @@ fn swap_three_different_pools_with_fees() {
         token3.address.clone(),
         2_000_000,
         Some(1_000),
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -477,6 +492,7 @@ fn swap_three_different_pools_with_fees() {
         token4.address.clone(),
         3_000_000,
         Some(1_000),
+        PoolType::Xyk,
     );
 
     // 4. swap with multihop
@@ -505,7 +521,7 @@ fn swap_three_different_pools_with_fees() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &operations, &None, &None, &10_000i128);
-    multihop.swap(&recipient, &operations, &None, &10_000i128);
+    multihop.swap(&recipient, &operations, &None, &10_000i128, &PoolType::Xyk);
 
     // we start swapping 10_000 tokens
 
@@ -550,7 +566,7 @@ fn swap_panics_with_no_operations() {
 
     // FIXM: Disable Referral struct
     // multihop.swap(&recipient, &None, &swap_vec, &None, &None, &50i128);
-    multihop.swap(&recipient, &swap_vec, &None, &50i128);
+    multihop.swap(&recipient, &swap_vec, &None, &50i128, &PoolType::Xyk);
 }
 
 #[test]
@@ -568,7 +584,7 @@ fn test_v_phx_vul_013_add_belief_price_for_every_swap() {
     let token4 = deploy_and_mint_tokens(&env, &admin, 10_000_000i128);
 
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -577,8 +593,9 @@ fn test_v_phx_vul_013_add_belief_price_for_every_swap() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -587,8 +604,9 @@ fn test_v_phx_vul_013_add_belief_price_for_every_swap() {
         token3.address.clone(),
         2_000_000,
         None,
+        PoolType::Xyk,
     );
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -597,6 +615,7 @@ fn test_v_phx_vul_013_add_belief_price_for_every_swap() {
         token4.address.clone(),
         3_000_000,
         None,
+        PoolType::Xyk,
     );
 
     let multihop = deploy_multihop_contract(&env, admin.clone(), &factory_client.address);
@@ -621,7 +640,7 @@ fn test_v_phx_vul_013_add_belief_price_for_every_swap() {
 
     let operations = vec![&env, swap1, swap2, swap3];
 
-    multihop.swap(&recipient, &operations, &None, &5_000i128);
+    multihop.swap(&recipient, &operations, &None, &5_000i128, &PoolType::Xyk);
 
     assert_eq!(
         token1.balance(&recipient),
@@ -649,7 +668,7 @@ fn test_swap_with_ask_asset_min_amount() {
 
     let factory_client = deploy_and_initialize_factory(&env, admin.clone());
 
-    deploy_and_initialize_lp(
+    deploy_and_initialize_pool(
         &env,
         &factory_client,
         admin.clone(),
@@ -658,6 +677,7 @@ fn test_swap_with_ask_asset_min_amount() {
         token2.address.clone(),
         1_000_000,
         None,
+        PoolType::Xyk,
     );
 
     let multihop = deploy_multihop_contract(&env, admin.clone(), &factory_client.address);
@@ -674,7 +694,7 @@ fn test_swap_with_ask_asset_min_amount() {
 
     let operations = vec![&env, swap1];
 
-    multihop.swap(&recipient, &operations, &None, &1_000);
+    multihop.swap(&recipient, &operations, &None, &1_000, &PoolType::Xyk);
 
     assert_eq!(token1.balance(&recipient), 4_000i128);
     assert_eq!(token2.balance(&recipient), 1_000i128);
@@ -685,5 +705,88 @@ fn test_swap_with_ask_asset_min_amount() {
         ask_asset_min_amount: Some(10_000),
     };
     let operations = vec![&env, greedy_swap];
-    multihop.swap(&recipient, &operations, &None, &1_000);
+    multihop.swap(&recipient, &operations, &None, &1_000, &PoolType::Xyk);
+}
+
+#[test]
+#[ignore = "fails with NewtonMethodFailed "]
+fn swap_three_equal_stable_pool() {
+    let env = Env::default();
+
+    let admin = Address::generate(&env);
+
+    env.mock_all_auths();
+    env.budget().reset_unlimited();
+
+    let token1 = deploy_and_mint_tokens(&env, &admin, 10_000_000i128);
+    let token2 = deploy_and_mint_tokens(&env, &admin, 10_000_000i128);
+    let token3 = deploy_and_mint_tokens(&env, &admin, 10_000_000i128);
+    let token4 = deploy_and_mint_tokens(&env, &admin, 10_000_000i128);
+
+    // 1. deploy factory
+    let factory_client = deploy_and_initialize_factory(&env, admin.clone());
+
+    deploy_and_initialize_pool(
+        &env,
+        &factory_client,
+        admin.clone(),
+        token1.address.clone(),
+        1_000_000,
+        token2.address.clone(),
+        1_000_000,
+        None,
+        PoolType::Stable,
+    );
+    deploy_and_initialize_pool(
+        &env,
+        &factory_client,
+        admin.clone(),
+        token2.address.clone(),
+        1_000_000,
+        token3.address.clone(),
+        1_000_000,
+        None,
+        PoolType::Stable,
+    );
+    deploy_and_initialize_pool(
+        &env,
+        &factory_client,
+        admin.clone(),
+        token3.address.clone(),
+        1_000_000,
+        token4.address.clone(),
+        1_000_000,
+        None,
+        PoolType::Stable,
+    );
+
+    // 4. swap with multihop
+    let multihop = deploy_multihop_contract(&env, admin.clone(), &factory_client.address);
+    let recipient = Address::generate(&env);
+    token1.mint(&recipient, &50i128);
+    assert_eq!(token1.balance(&recipient), 50i128);
+    assert_eq!(token4.balance(&recipient), 0i128);
+
+    let swap1 = Swap {
+        offer_asset: token1.address.clone(),
+        ask_asset: token2.address.clone(),
+        ask_asset_min_amount: None::<i128>,
+    };
+    let swap2 = Swap {
+        offer_asset: token2.address.clone(),
+        ask_asset: token3.address.clone(),
+        ask_asset_min_amount: None::<i128>,
+    };
+    let swap3 = Swap {
+        offer_asset: token3.address.clone(),
+        ask_asset: token4.address.clone(),
+        ask_asset_min_amount: None::<i128>,
+    };
+
+    let operations = vec![&env, swap1, swap2, swap3];
+
+    multihop.swap(&recipient, &operations, &None, &50i128, &PoolType::Stable);
+
+    assert_eq!(token1.balance(&recipient), 0i128);
+    assert_eq!(token4.balance(&recipient), 50i128);
 }